refactor: split Dashboard.tsx (2,051 lines) into 14 modular components
Extracted 11 page components, shared UI library, and types into
frontend/src/app/components/dashboard/ directory. Dashboard.tsx is
now 284 lines (shell + layout) with each page component under 200 lines.
